able to save and load nature data
