+++ vite-project/src/pages/PokemonTeamBuilder.tsx
@@ -5,6 +5,7 @@ import { PokemonTeam } from '../models/PokemonTeamsInterface';
 import TeraTypeSelector from '../components/pokemonComponents/TeraTypeSelector';
 import AbilitiesSelector from '../components/pokemonComponents/AbilitiesSelector';
 import MoveSlotSelector from '../components/pokemonComponents/MoveSlotSelector';
+import {PokemonNatureSelector} from '../components/pokemonComponents/PokemonNature';
 // import SpriteSelector from '../components/pokemonComponents/SpriteSelector';
 
 const PokemonTeamBuilder: React.FC = () => {
@@ -23,6 +24,7 @@ const PokemonTeamBuilder: React.FC = () => {
     const [move2, setMove2] = useState('')
     const [move3, setMove3] = useState('')
     const [move4, setMove4] = useState('')
+    const [currentNature, setCurrentNature] = useState('')
 
     useEffect(() =>{
         const savedTeam = localStorage.getItem('savedPokemonTeam')
@@ -40,6 +42,14 @@ const PokemonTeamBuilder: React.FC = () => {
         } catch (error) {
             console.log("Can't find pokemon", error)
         }
+        setSelectedTeraType('')
+        setAbility('')
+        setMove1('')
+        setMove2('')
+        setMove2('')
+        setMove2('')
+        setCurrentNature('')
+
     }
     
     const handleSpriteSelect = (spriteUrl :string) => {
@@ -58,15 +68,15 @@ const PokemonTeamBuilder: React.FC = () => {
                 move_1: move1,
                 move_2: move2,
                 move_3: move3,
-                move_4: move4
+                move_4: move4,
+                nature: currentNature
             }
             if (editMode && savedPokemonTeam) {
                 const updatedPokemon = savedPokemonTeam?.pokemons.map((pokemon) =>
                     pokemon.id === editedPokemon.id ? 
-                    {...pokemon, pokemon: editedPokemon, id: `pokemon-${pokemonData.name}-
-                        ${savedPokemonTeam ? savedPokemonTeam.pokemons.length + 1 : 1}`, 
+                    {...pokemon, pokemon: editedPokemon, id: `pokemon-${pokemonData.name}-${savedPokemonTeam ? savedPokemonTeam.pokemons.length + 1 : 1}`, 
                         data: pokemonData, sprite: selectedSprite, teraType: selectedTeraType, ability : ability,
-                        move_1: move1, move_2: move2, move_3: move3, move_4: move4
+                        move_1: move1, move_2: move2, move_3: move3, move_4: move4, nature: currentNature
                         } : pokemon)
                     const updatedTeam = {
                         ...savedPokemonTeam,
@@ -84,7 +94,8 @@ const PokemonTeamBuilder: React.FC = () => {
                     move_1: move1,
                     move_2: move2,
                     move_3: move3,
-                    move_4: move4
+                    move_4: move4,
+                    nature: currentNature
                 }
                 const updatedTeam = savedPokemonTeam 
                 ? {
@@ -103,7 +114,8 @@ const PokemonTeamBuilder: React.FC = () => {
     }
 
     const loadPokemonOnClick = (selectedPokemon: { id: string, data: Pokemon, sprite: string, 
-        teraType: string, ability :string, move_1: string, move_2: string, move_3: string, move_4: string
+        teraType: string, ability :string, move_1: string, move_2: string, move_3: string, move_4: string,
+        nature: string
     }) => {
         setPokemonID(selectedPokemon.id)
         setPokemonData(selectedPokemon.data)
@@ -114,6 +126,7 @@ const PokemonTeamBuilder: React.FC = () => {
         setMove2(selectedPokemon.move_2)
         setMove2(selectedPokemon.move_3)
         setMove2(selectedPokemon.move_4)
+        setCurrentNature(selectedPokemon.nature)
         setEditMode(true);
     }
 
@@ -157,6 +170,7 @@ const PokemonTeamBuilder: React.FC = () => {
                     </div>
                     <button onClick={handleSavePokemon}>Save</button>
                     <TeraTypeSelector setSelectedTeraType={setSelectedTeraType} selectedTeraType={selectedTeraType}></TeraTypeSelector>
+                    <PokemonNatureSelector setCurrentNature={setCurrentNature} currentNature={currentNature}></PokemonNatureSelector>
                     <AbilitiesSelector abilityUrls={pokemonData.abilities.map(ability => ability.ability.url)} selectedAbility={ability} setSelectedAbility={setAbility}></AbilitiesSelector>
                     <MoveSlotSelector moveNames={pokemonData.moves.map((move) => move.move.name)} selectedMove={move1} setSelectedMove={setMove1}></MoveSlotSelector>
                     <MoveSlotSelector moveNames={pokemonData.moves.map((move) => move.move.name)} selectedMove={move2} setSelectedMove={setMove2}></MoveSlotSelector>
